pyproject.toml
[tool.poetry]
name = "dls-simpsons-classification"
version = "0.1.0"
description = "Classification of the characters of the series “The Simpsons”"

license = "MIT"

authors = [
    "Ruslan Mukhametshin <rusmux21@gmail.com>"
]

readme = "README.md"
homepage = ""


[tool.poetry.dependencies]
python = ">=3.8, <3.11"
python-dotenv = "^0.5.1"
jupyterlab = "^3.5.0"
ipywidgets = "^8.0.2"
jupyter-contrib-nbextensions = "^0.5.1"

click = "^8.1.3"
pydantic = "^1.10.2"

numpy = "^1.23.4"
pandas = "^1.5.1"
Pillow = "^9.2.0"

scikit-learn = "^1.1.3"
torch = "~1.12.1"
torchvision = "^0.13.1"
torchinfo = "^1.7.1"
tensorboard = "^2.10.1"
torch-tb-profiler = "^0.4.0"

plotly = "^5.10.0"
kaleido = "0.2.1"
custom-plotly-templates = "^0.1.1"
tqdm = "^4.64.1"

kaggle = "^1.5.12"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^22.10.0"}
isort = "^5.10.1"
pylint = "^2.15.5"
flake8 = "^5.0.4"
mypy = "^0.982"
bandit = "^1.7.4"
nbqa = "^1.5.3"

[tool.black]
line-length = 110

[tool.pylint.MASTER]
max-line-length=110
disable = [
   "missing-module-docstring",
   "missing-class-docstring",
   "missing-function-docstring",
   "logging-fstring-interpolation",
]
generated-members= ["numpy.*", "torch.*"]
good-names = ["i", "x", "fc", "X_batch"]
extension-pkg-whitelist = "pydantic"

[tool.mypy]
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 110

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
